tests/test_fail.html

<!DOCTYPE html>
<html>
  <head>
    <title>HTML Test: Forms date state</title>
    <link rel='author' title='your mame' href='mailto:youremail@address.com'>
    <link rel='help' href='http://www.w3.org/html/wg/drafts/html/CR/forms.html#date-state-(type=date)'>
    <script src='testharness.js'></script>
    <script src='testharnessreport.js'></script>
  </head>
  
  <body>
    <div id="myDiv" style="display: none">
      <input id="invalid_value" type="date" value="2013-12-07">
    </div>
    <div id="log"></div>    
    <script>
        test(function() {
          assert_equals(document.getElementById("invalid_value").value, "");
        },
        "User agents must not allow the user to set the value to"
        + " a non-empty string that is not a valid date string.");
    </script>
  </body>
</html>
